+++ footer.py
@@ -1,0 +1,7 @@
+def get_footer():
+    return '''
+    <footer>
+        <hr>
+        <p>Chess Opening Explorer by Jimmy Rustles</p>
+    </footer>
+    '''
