+++ requirements.txt
@@ -1,0 +1,4 @@
+Flask==2.3.2
+beautifulsoup4==4.12.2
+html5lib==1.1
+python-chess==1.999
